# openapi.json
{
  "components": {
    "schemas": {
...
      "ContentDTO": {
        "additionalProperties": false,
        "properties": {
          "archived_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Archived At"
          },
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_pinned": {
            "default": false,
            "title": "Is Pinned",
            "type": "boolean"
          },
          "owner_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "owner_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Type"
          },
          "pin_rank": {
            "default": 0,
            "title": "Pin Rank",
            "type": "integer"
          },
          "publish_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Publish At"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "schedule_version": {
            "default": 0,
            "title": "Schedule Version",
            "type": "integer"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "string"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type_name": {
            "title": "Type Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "version": {
            "default": 1,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "type_name",
          "schema_version",
          "title",
          "slug",
          "status",
          "data",
          "created_at",
          "updated_at"
        ],
        "title": "ContentDTO",
        "type": "object"
      },
      "ContentPageDTO": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContentDTO"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "size": {
            "title": "Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "size"
        ],
        "title": "ContentPageDTO",
        "type": "object"
      },
      "CreateContentInput": {
        "additionalProperties": false,
        "properties": {
          "body": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Body"
          },
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "excerpt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Excerpt"
          },
          "owner_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Owner Id"
          },
          "slug": {
            "title": "Slug",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type_name": {
            "title": "Type Name",
            "type": "string"
          }
        },
        "required": [
          "type_name",
          "title",
          "slug"
        ],
        "title": "CreateContentInput",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "aiya-cms (cms)",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/content": {
      "get": {
        "operationId": "list_content_api_v1_admin_content_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "type_name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Type Name"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sort"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentPageDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      },
      "post": {
        "operationId": "create_content_api_v1_admin_content_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateContentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}": {
      "get": {
        "operationId": "get_content_api_v1_admin_content__content_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      },
      "patch": {
        "operationId": "update_content_api_v1_admin_content__content_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateContentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/archive": {
      "post": {
        "operationId": "archive_content_api_v1_admin_content__content_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Archive Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/pin": {
      "post": {
        "operationId": "set_pin_api_v1_admin_content__content_id__pin_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetContentPinInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Set Pin",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/publish": {
      "post": {
        "operationId": "publish_content_api_v1_admin_content__content_id__publish_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Publish Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/purge": {
      "post": {
        "operationId": "purge_content_api_v1_admin_content__content_id__purge_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": false,
              "title": "Dry Run",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurgeResultDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Purge Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/references": {
      "get": {
        "operationId": "list_references_api_v1_admin_content__content_id__references_get",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ReferenceDTO"
                  },
                  "title": "Response List References Api V1 Admin Content  Content Id  References Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List References",
        "tags": [
          "admin",
          "admin-content"
        ]
      },
      "put": {
        "operationId": "replace_references_api_v1_admin_content__content_id__references_put",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceReferencesInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Replace References",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/reject": {
      "post": {
        "operationId": "reject_content_api_v1_admin_content__content_id__reject_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reject Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/restore": {
      "post": {
        "operationId": "restore_content_api_v1_admin_content__content_id__restore_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Restore Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/schedule": {
      "post": {
        "operationId": "schedule_content_api_v1_admin_content__content_id__schedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleContentInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Schedule Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/submit": {
      "post": {
        "operationId": "submit_content_api_v1_admin_content__content_id__submit_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
    "/api/v1/admin/content/{content_id}/unschedule": {
      "post": {
        "operationId": "unschedule_content_api_v1_admin_content__content_id__unschedule_post",
        "parameters": [
          {
            "in": "path",
            "name": "content_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Content Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentDTO"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Unschedule Content",
        "tags": [
          "admin",
          "admin-content"
        ]
      }
    },
...
  },
...
}